requirements.txt
# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Data processing
pyarrow>=10.0.0
fastparquet>=0.8.0

# Hyperparameter optimization
optuna>=3.0.0

# Configuration
PyYAML>=6.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Code quality
flake8>=5.0.0
black>=22.0.0
isort>=5.0.0

# Utilities
tqdm>=4.60.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Optional: Jupyter support
jupyter>=1.0.0
ipykernel>=6.0.0

# Optional: Weights & Biases for experiment tracking
# wandb>=0.12.0

# Optional: TensorBoard for visualization
# tensorboard>=2.8.0
